+++ FieldtypeSnipWireTaxSelector/FieldtypeSnipWireTaxSelector.module.php
@@ -17,7 +17,7 @@ namespace ProcessWire;
  *
  */
 
-require_once dirname(dirname(__FILE__)) . '/helpers/Taxes.php';
+require_once dirname(__DIR__) . '/helpers/Taxes.php';
 
 use SnipWire\Helpers\Taxes;