+++ MarkupSnipWire/MarkupSnipWire.module.php
@@ -13,7 +13,7 @@ namespace ProcessWire;
  *
  */
  
-require_once dirname(dirname(__FILE__)) . '/helpers/CurrencyFormat.php';
+require_once dirname(__DIR__) . '/helpers/CurrencyFormat.php';
 
 use SnipWire\Helpers\CurrencyFormat;