+++ ProcessSnipWire/ProcessSnipWire.module.php
@@ -13,17 +13,17 @@ namespace ProcessWire;
  *
  */
 
-require_once dirname(dirname(__FILE__)) . '/helpers/Functions.php';
-require_once dirname(dirname(__FILE__)) . '/helpers/CurrencyFormat.php';
-require_once dirname(dirname(__FILE__)) . '/helpers/Countries.php';
-require_once dirname(dirname(__FILE__)) . '/installer/ExtendedInstaller.php';
-require_once dirname(__FILE__) . '/sections/Dashboard.php';
-require_once dirname(__FILE__) . '/sections/Orders.php';
-require_once dirname(__FILE__) . '/sections/Subscriptions.php';
-require_once dirname(__FILE__) . '/sections/AbandonedCarts.php';
-require_once dirname(__FILE__) . '/sections/Customers.php';
-require_once dirname(__FILE__) . '/sections/Products.php';
-require_once dirname(__FILE__) . '/sections/Discounts.php';
+require_once dirname(__DIR__) . '/helpers/Functions.php';
+require_once dirname(__DIR__) . '/helpers/CurrencyFormat.php';
+require_once dirname(__DIR__) . '/helpers/Countries.php';
+require_once dirname(__DIR__) . '/installer/ExtendedInstaller.php';
+require_once __DIR__ . '/sections/Dashboard.php';
+require_once __DIR__ . '/sections/Orders.php';
+require_once __DIR__ . '/sections/Subscriptions.php';
+require_once __DIR__ . '/sections/AbandonedCarts.php';
+require_once __DIR__ . '/sections/Customers.php';
+require_once __DIR__ . '/sections/Products.php';
+require_once __DIR__ . '/sections/Discounts.php';
 
 use SnipWire\Installer\ExtendedInstaller;
 use SnipWire\Helpers\CurrencyFormat;