+++ SnipWireConfig.php
@@ -13,9 +13,9 @@ namespace ProcessWire;
  *
  */
 
-require_once dirname(__FILE__) . '/helpers/Functions.php';
-require_once dirname(__FILE__) . '/helpers/CurrencyFormat.php';
-require_once dirname(__FILE__) . '/helpers/Taxes.php';
+require_once __DIR__ . '/helpers/Functions.php';
+require_once __DIR__ . '/helpers/CurrencyFormat.php';
+require_once __DIR__ . '/helpers/Taxes.php';
 
 use SnipWire\Helpers\CurrencyFormat;
 use SnipWire\Helpers\Taxes;